+++ CHANGELOG.md
@@ -5,7 +5,18 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.5.12] - 2023-02-27
+## [0.5.13] - 2023-02-28
+
+### Added
+
+- Options to date proxy and boolean proxy.
+
+### Fixed
+
+- Date proxy now working properly.
+- `SuperDebug` date output now handles invalid dates.
+
+## [0.5.12] - 2023-02-28
 
 ### Fixed